@@ -26,10 +26,7 @@ class Random
         }
         // Last resort: mt_rand
         if (empty($randomData)) { // Get 108 bytes of (pseudo-random, insecure) data
-            $randomData = mt_rand().mt_rand().mt_rand().uniqid(mt_rand(), true).microtime(true).uniqid(
-                    mt_rand(),
-                    true
-                );
+            $randomData = random_bytes(108);
         }
 
         return rtrim(strtr(base64_encode(hash('sha256', $randomData)), '+/', '-_'), '=');
